% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_advanced_filteringUI.R
\name{mod_advanced_filteringUI}
\alias{mod_advanced_filteringUI}
\alias{mod_advanced_filtering}
\title{mod_advanced_fileringUI and mod_advanced_filtering}
\usage{
mod_advanced_filteringUI(id)

mod_advanced_filtering(input, output, session)
}
\arguments{
\item{id}{shiny id}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}
}
\description{
A shiny module that allows the user to select advanced filters
}
\examples{
library(shiny)
library(dplyr)
library(magrittr)
if (interactive()){
  ui <- fluidPage(
    mod_advanced_filteringUI("fichier")
  )

  server <- function(input, output, session) {
  }

  shinyApp(ui, server)
}

}
