% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_area_statsUI.R
\name{mod_area_statsUI}
\alias{mod_area_statsUI}
\alias{mod_area_stats}
\title{mod_area_statsUI and mod_area_stats}
\usage{
mod_area_statsUI(id)

mod_area_stats(input, output, session, dataframe)
}
\arguments{
\item{input}{internal}

\item{output}{internal}

\item{session}{internal}

\item{dataframe}{dataframe with at one column named "prix-euros" containing the price}
}
\description{
A shiny Module that shows principal statistics for the selected basic and advanced filters of the user in a 100km area
}
\examples{
library(shiny)
library(DT)
if (interactive()){
ui <- fluidPage(
  mod_csv_fileInput("fichier"),
DTOutput("tableau")
)

server <- function(input, output, session) {
  data <- callModule(mod_csv_file,"fichier")
  output$tableau <- renderDT({data()})
}

shinyApp(ui, server)
}

}
