% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_basic_filteringUI.R
\name{mod_basic_filteringUI}
\alias{mod_basic_filteringUI}
\alias{mod_basic_filtering}
\title{mod_basic_fileringUI and mod_basic_filtering}
\usage{
mod_basic_filteringUI(id)

mod_basic_filtering(input, output, session)
}
\arguments{
\item{id}{shiny id}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}
}
\description{
A shiny module that allows the user to select basic filters
}
\examples{
library(shiny)
library(dplyr)
library(magrittr)
library(utils)
if (interactive()){
  ui <- fluidPage(
    mod_basic_filteringUI("fichier")
  )

  server <- function(input, output, session) {
  }

  shinyApp(ui, server)
}

}
