% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_map_countryUI.R
\name{mod_map_countryUI}
\alias{mod_map_countryUI}
\alias{mod_map_country}
\title{mod_map_countryUI and mod_map_country}
\usage{
mod_map_countryUI(id)

mod_map_country(input, output, session, dataframe)
}
\arguments{
\item{id}{shiny id}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}

\item{dataframe}{dataframe with columns named "prix-euros", "nom_commune", "carrosserie", "transmission", "brand", "date", "energie", "nb_places", "kilometrage_km" and "nb_portes"}
}
\description{
A shiny module that displays a map of the results in the whole country by region
}
\examples{
"No example to display"

}
