% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_regressionUI.R
\name{mod_regressionUI}
\alias{mod_regressionUI}
\alias{mod_regression}
\title{mod_regressionUI and mod_regression}
\usage{
mod_regressionUI(id)

mod_regression(input, output, session, dataframe)
}
\arguments{
\item{id}{shiny id}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}

\item{dataframe}{dataframe with columns named "prix-euros", "nom_commune", "carrosserie", "transmission", "brand", "date", "energie", "nb_places", "kilometrage_km" and "nb_portes"}
}
\description{
A shiny Module that performs a regression of the price with respect to other parameters and displays the results for a specific type of car
}
\examples{
"No example to display"

}
