% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_table_areaUI.R
\name{mod_table_areaUI}
\alias{mod_table_areaUI}
\alias{mod_table_area}
\title{mod_table_areaUI and mod_table_area}
\usage{
mod_table_areaUI(id)

mod_table_area(input, output, session, dataframe)
}
\arguments{
\item{id}{shiny id}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}

\item{dataframe}{dataframe with columns named "prix-euros", "latitude", "longitude", "nom_commune", "carrosserie", "transmission", "brand", "date", "energie", "nb_places", "kilometrage_km" and "nb_portes"}
}
\description{
A shiny module that displays a table of the results for a 100km radius around the user's chosen city
}
\examples{
"No example to display"

}
