% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_table_countryUI.R
\name{mod_table_countryUI}
\alias{mod_table_countryUI}
\alias{mod_table_country}
\title{mod_table_countryUI and mod_table_country}
\usage{
mod_table_countryUI(id)

mod_table_country(input, output, session, dataframe)
}
\arguments{
\item{id}{shiny id}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}

\item{dataframe}{dataframe with columns named "prix-euros", "nom_commune", "carrosserie", "transmission", "brand", "date", "energie", "nb_places", "kilometrage_km" and "nb_portes"}
}
\description{
A shiny module that displays a table of the results in the whole country
}
\examples{
"No example to display"

}
