% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_advanced_filteringUI.R
\name{mod_advanced_filteringUI}
\alias{mod_advanced_filteringUI}
\alias{mod_advanced_filtering}
\title{mod_advanced_fileringUI and mod_advanced_filtering}
\usage{
mod_advanced_filteringUI(id)

mod_advanced_filtering(input, output, session)
}
\arguments{
\item{input}{internal}

\item{output}{internal}

\item{session}{internal}
}
\description{
A shiny Module that allows the user to select the advanced filters
}
\examples{
library(shiny)
library(DT)
if (interactive()){
ui <- fluidPage(
  mod_csv_fileInput("fichier"),
DTOutput("tableau")
)

server <- function(input, output, session) {
  data <- callModule(mod_csv_file,"fichier")
  output$tableau <- renderDT({data()})
}

shinyApp(ui, server)
}

}
