% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_area_statsUI.R
\name{mod_area_statsUI}
\alias{mod_area_statsUI}
\alias{mod_area_stats}
\title{mod_area_statsUI and mod_area_stats}
\usage{
mod_area_statsUI(id)

mod_area_stats(input, output, session, dataframe, city, tcarrosserie)
}
\arguments{
\item{id}{shiny id}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}

\item{dataframe}{dataframe with columns named "prix-euros", "latitude", "longitude", "nom_commune", "carrosserie", "transmission", "brand", "date", "energie", "nb_places", "kilometrage_km" and "nb_portes"}
}
\description{
A shiny module that displays principal statistics for the basic and advanced filters selected by the user in a 100km radius
}
\examples{
library(shiny)
library(dplyr)
library(magrittr)
if (interactive()){
  ui <- fluidPage(
    mod_country_statsUI("fichier")
  )

  server <- function(input, output, session) {
    callModule(mod_country_stats, "fichier", cardata)
  }

  shinyApp(ui, server)
}

}
