% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_country_statsUI.R
\name{mod_country_statsUI}
\alias{mod_country_statsUI}
\alias{mod_country_stats}
\title{mod_country_statsUI and mod_country_stats}
\usage{
mod_country_statsUI(id)

mod_country_stats(input, output, session, dataframe)
}
\arguments{
\item{id}{shiny id}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}

\item{dataframe}{dataframe with columns named "prix-euros", "nom_commune", "carrosserie", "transmission", "brand", "date", "energie", "nb_places", "kilometrage_km" and "nb_portes"}
}
\description{
A shiny module that displays principal statistics for the basic and advanced filters selected by the user in the whole country
}
\examples{
library(shiny)
library(DT)
if (interactive()){
ui <- fluidPage(
  mod_csv_fileInput("fichier"),
DTOutput("tableau")
)

server <- function(input, output, session) {
  data <- callModule(mod_csv_file,"fichier")
  output$tableau <- renderDT({data()})
}

shinyApp(ui, server)
}

}
