% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_map_countryUI.R
\name{mod_map_countryUI}
\alias{mod_map_countryUI}
\alias{mod_map_country}
\title{mod_map_countryUI and mod_map_country}
\usage{
mod_map_countryUI(id)

mod_map_country(input, output, session, dataframe)
}
\arguments{
\item{input}{internal}

\item{output}{internal}

\item{session}{internal}

\item{dataframe}{dataframe with at one column named "prix-euros" containing the price}
}
\description{
A shiny Module that shows principal statistics for the selected basic and advanced filters of the user in a 100km area
}
\examples{
library(shiny)
library(DT)
if (interactive()){
ui <- fluidPage(
  mod_csv_fileInput("fichier"),
DTOutput("tableau")
)

server <- function(input, output, session) {
  data <- callModule(mod_csv_file,"fichier")
  output$tableau <- renderDT({data()})
}

shinyApp(ui, server)
}

}
